LayoutTests/accessibility/ax-thread-text-apis/attributed-string/heading-level-expected.txt

This test ensures we represent AXHeadingLevel in attributed strings.

PASS: fooText.attributedStringForTextMarkerRangeContainsAttribute('AXHeadingLevel', markerRange) === true
PASS: barText.attributedStringForTextMarkerRangeContainsAttribute('AXHeadingLevel', markerRange) === true
PASS: AXHeadingLevel is not present after dynamic role attribute change.

PASS successfullyParsed is true

TEST COMPLETE
Foo

Bar
